+++ CHANGELOG.md
@@ -2,9 +2,9 @@
 
 ## Dev
 
+- Added resolvers
+- Added Resolver class. Container is refactored to use Resolver class
 - Added multiple containers support(instead of scopes)
-- Added Resolver class
-- Container is refactored to use Resolver class
 
 ## 1.0.1